# swagger.yaml
openapi: 3.0.0
paths:
...
  /profiles:
    get:
      operationId: ProfilesController_findAll
      summary: Получить все профили
      parameters: []
      responses:
        '200':
          description: Профили успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfileResponseBodyOK'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: &ref_1
        - profiles
      security: &ref_3
        - &ref_2
          bearer: []
  /profiles/me:
    get:
      operationId: ProfilesController_findProfileByToken
      summary: Получить текущий профиль
      parameters:
        - name: authorization
          required: true
          in: header
          description: Access токен
          schema:
            type: string
      responses:
        '200':
          description: Профиль успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponseBodyOK'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_1
      security:
        - *ref_2
        - bearer: []
  /profiles/{id}:
    get:
      operationId: ProfilesController_findOne
      summary: Получить профиль по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор профиля
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Профиль успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponseBodyOK'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_1
      security: *ref_3
    patch:
      operationId: ProfilesController_update
      summary: Обновить данные профиля по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор профиля
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Профиль успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponseBodyOK'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_1
      security: *ref_3
    delete:
      operationId: ProfilesController_remove
      summary: Удалить профиль по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор профиля
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Профиль успешно удален
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponseBodyOK'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_1
      security: *ref_3
  /profiles/{id}/accounts:
    get:
      operationId: ProfilesController_findAccountByProfileId
      summary: Получить все аккаунты пользователя по id профиля
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор профиля
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Аккаунты профиля успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SingleAccountResponseBodyOK'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_1
      security: *ref_3
...
  /platforms:
    post:
      operationId: PlatformController_create
      summary: Создать новую платформу
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlatformDto'
      responses:
        '201':
          description: Платформа успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
      tags: &ref_10
        - platforms
      security: &ref_11
        - bearer: []
    get:
      operationId: PlatformController_findAll
      summary: Получить все платформы
      parameters: []
      responses:
        '200':
          description: Платформы успешно получены
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Platform'
        '403':
          description: Отказ в доступе
      tags: *ref_10
      security: *ref_11
  /platforms/{id}:
    get:
      operationId: PlatformController_findOne
      summary: Получить платформу по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентиификатор платформы
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Платформа успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_10
      security: *ref_11
    patch:
      operationId: PlatformController_update
      summary: Обновить данные о платформе по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор платформы
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlatformDto'
      responses:
        '200':
          description: Платформа успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_10
      security: *ref_11
    delete:
      operationId: PlatformController_remove
      summary: Удалить платформу по id
      parameters:
        - name: id
          required: true
          in: path
          description: Идентификатор платформы
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      responses:
        '200':
          description: Платформа успешно удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Platform'
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_10
      security: *ref_11
...
info:
  title: API BotKits
  description: Ручки для команды frontend
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UserProfileResponseBodyOK:
      type: object
      properties:
        _id:
          type: string
          example: 65196b9715b55bd51b039144
          description: ' '
        username:
          type: string
          example: test
          description: ' '
        phone:
          type: string
          example: '+79999999999'
          description: ' '
        avatar:
          type: string
          example: https://i.pravatar.cc/300
          description: ' '
        balance:
          type: number
          example: 0
          description: ' '
        accounts:
          type: array
          example:
            - 65196b9715b55bd51b039144
        success:
          type: boolean
          example: true
      required:
        - _id
        - username
        - phone
        - avatar
        - balance
        - accounts
    SingleAccountResponseBodyOK:
      type: object
      properties:
        _id:
          type: string
          example: 6516ab8f2c612e8cce53e45f
        type:
          type: string
          example: local
        role:
          type: string
          example: user
        credentials:
          type: object
          example:
            email: test@mail.ru
            accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
            refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        profile:
          type: string
          example: 6516ab8f2c612e8cce53e45d
        success:
          type: boolean
          example: true
      required:
        - credentials
    UpdateProfileDto:
      type: object
      properties: {}
...
    FormFields:
      type: object
      properties:
        name:
          type: boolean
          example: true
        pages:
          type: boolean
          example: true
        accessKey:
          type: boolean
          example: false
        url:
          type: boolean
          example: true
      required:
        - name
        - pages
        - accessKey
        - url
    CreatePlatformDto:
      type: object
      properties:
        icon:
          type: string
          example: >-
            https://cdn.icon-icons.com/icons2/1233/PNG/512/1492718766-vk_83600.png
        title:
          type: string
          example: VK
        formFields:
          $ref: '#/components/schemas/FormFields'
      required:
        - icon
        - title
        - formFields
    Platform:
      type: object
      properties:
        icon:
          type: string
          example: >-
            https://cdn.icon-icons.com/icons2/1233/PNG/512/1492718766-vk_83600.png
        title:
          type: string
          example: VK
        formFields:
          example:
            name: true
            pages: true
            accessKey: false
            url: true
          allOf:
            - $ref: '#/components/schemas/FormFields'
      required:
        - icon
        - title
        - formFields
    UpdatePlatformDto:
      type: object
      properties: {}
...